Pattern 2b: Remove redundant u32/s32 casts in comparison contexts
Extended redundant cast removal to u32/s32 types, restricted to safe
comparison and loop-condition contexts to avoid breaking pointer-to-int
conversions. Removed 4432 redundant casts across 18 files including
colosseum_battle.c (1439), gs_field_world.c (890), trainer.c (565),
ui_core.c (453), pokemon.c (437).

+++ src/game/colosseum_event.c
@@ -8143,7 +8143,7 @@ void fn_8020E124(void) {
     r4 = lbl_80478F00;
     r7 = r3 & 0xFFFF;
     r6 = *(u32*)((u8*)r4 + 0x0);
-    if ((u32)r7 > (u32)r6) {
+    if (r7 > r6) {
         r4 = 0x0;
     } else {
 
@@ -8151,13 +8151,13 @@ void fn_8020E124(void) {
         /* clrlslwi r0, r3, 16, 3 */;
         r4 = r4 + r0;
     }
-    if ((u32)r4 == (u32)0x0) {
+    if (r4 == (u32)0x0) {
         r5 = 0x0;
     } else {
 
         r5 = *(u8*)((u8*)r4 + 0x0);
     }
-    if ((u32)r7 > (u32)r6) {
+    if (r7 > r6) {
         r4 = 0x0;
     } else {
 
@@ -8166,7 +8166,7 @@ void fn_8020E124(void) {
         r4 = r4 + r0;
     }
     r3 = r5 & 0xFF;
-    if ((u32)r4 == (u32)0x0) {
+    if (r4 == (u32)0x0) {
         r0 = 0x0;
     } else {
